risk_engine/requirements.txt
pandas==2.1.4
requests==2.31.0
scikit-learn==1.3.2
numpy==1.24.4
python-dotenv==1.0.0
schedule==1.2.0
streamlit==1.29.0
plotly==5.17.0
sqlalchemy==2.0.23